Add list item option action, fix bugs, refactor

+++ app/src/main/java/com/geckour/q/domain/model/PlayerState.kt
@@ -3,6 +3,6 @@ package com.geckour.q.domain.model
 data class PlayerState(
         val queue: List<Song>,
         val currentPosition: Int,
-        val progress: Float,
+        val progress: Long,
         val playWhenReady: Boolean
 )
